third_party/blink/web_tests/fast/forms/suggestion-picker/month-suggestion-picker-min-max-attribute.html

<!DOCTYPE html>
<meta name=fuzzy content="maxDifference=0-3; totalPixels=0-1000">
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/testdriver.js"></script>
<script src="../../../resources/testdriver-vendor.js"></script>
<script src="../resources/picker-common.js"></script>
<script src="resources/suggestion-picker-common.js"></script>

<body style="background-color: #bbbbbb;">
<input type=month value="2012-12" min="2012-02" max="2012-05" list=suggestions>
<datalist id=suggestions>
  <option>2012-01</option>
  <option>2012-02</option>
  <option>2012-03</option>
  <option>2012-04</option>
  <option>2012-05</option>
  <option>2012-06</option>
  <option>2012-07</option>
  <option>2012-08</option>
  <option>2012-09</option>
  <option>2012-10</option>
</datalist>

<script>
window.onload = () => {
  const element = document.querySelector('input');
  promise_test(() => {
    return openPicker(element)
    .then(() => {
      internals.pagePopupWindow.focus();

      assert_equals(entryValues().length, 5);
      assert_equals(entryValues()[0], '2012-02');
      assert_equals(entryValues()[1], '2012-03');
      assert_equals(entryValues()[2], '2012-04');
      assert_equals(entryValues()[3], '2012-05');
      assert_equals(entryValues()[4], '@openCalendarPicker');

      return attemptToClosePicker(element);
    });
  },'min/max attributes filter out suggestions');
};
</script>
